requirements.txt
openai>=1.0.0,<2.0.0
streamlit
streamlit-chat
python-dotenv